+++ paradise_server/src/main/java/com/iaa/paradise_server/Service/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package com.iaa.paradise_server;
+package com.iaa.paradise_server.Service;
 
 import com.iaa.paradise_server.Entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
